[Fix] #442 - 한 줄 소개 기획 변경 사항 반영
