+++ SOPT-iOS/Projects/Core/Sources/Literals/StringLiterals.swift
@@ -263,7 +263,7 @@ public struct I18N {
     
     public struct Soptlog {
         public static let editProfile = "프로필 수정"
-        public static let enrollIntroduce = "한 줄 소개 등록하기 >"
+        public static let enrollIntroduce = "프로필 수정에서 한 줄 소개 등록해보세요!"
     }
     
     public struct Attendance {
