--- SOPT-iOS/Projects/Features/SoptlogFeature/Sources/Empty.swift
@@ -1,8 +1,0 @@
-//
-//  Empty.swift
-//  ProjectDescriptionHelpers
-//
-//  Created by 강윤서 on 11/25/24.
-//
-
-import Foundation
